backend/requirements.txt
Flask>=2.2,<3.0
Flask-Cors>=3.0
python-dotenv>=0.21
pymongo>=4.0
dnspython>=2.0
